fix a bug in the register_class decorator
because of the way the decorators work, parent classes' registries were
being updated by daughter classes in a way that made the parent
class's methods uncallable (they were updated so that the parent classes'
methods came to expect an instantiation of the daughter class as self).
This fixes that by making sure that methods are not added to the
methodRegistry more than once.