minimal changes to make testColumnOrigins.py use pytest
this required a change in how the unit tests which verify that
column origins point to the correct mixin parse the string
representations of methods.  Running the unittest with python
causes the mixins to appear to belong to '__main__'.  Running
with py.test causes the mixins to appear to belong to their
proper classes.