Rewrite tests to use temporary scratch space
When pytest runs jobs in parallel, multiple tests will be using the
scratch directories. This caused many tests to fail as they were
racing each other to database files and SQLite reported many
problems.

This patch rewrites the tests to use mkdtemp(). Two of the tests,
testCompoundInstanceCatalog and testCatalogDBObject, use the scratch
space in class attributes. This leads to somewhat convoluted logic
where the temp directory is allocated at setup_module(), which
runs after the code has been imported (so the temporary directory
can not be known to the class immediately), and the class has
to have an __init__ call that assigns the database location.

One improvement would be to assign the scratch space based on
the name of the class, but for now I've used "scratchSpace-"
as the prefix. This has been tested with multiple pytest
processes and seems to be working well.