jsonize test varParamStr by hand
So that we can control the length of the string in dtype.
There was a test failure where dtype was set on a shorter
version of varParamStr, causing longer varParamStr to be
incorrectly loaded by json.

This has exposed a problem with DBObject's automatic dtype
detection.  Strings must be of a uniform length, or there
is a risk that dtype will be improperly set.  I think this
is a reasonable risk.  Users really should be passing in
dtypes by hand, anyway.